backend/requirements.txt
aiodns==3.1.1
aiosmtpd==1.4.5
amqp==5.2.0
asgiref==3.7.2
atpublic==4.0
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
deprecation==2.1.0
Django==5.0.3
django-celery==3.1.17
django-cors-headers==4.3.1
django-email-verification==0.3.3
django-filter==24.3
django-jazzmin==3.0.0
django-use-email-as-username==1.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
getmac==0.9.4
kombu==5.4.0
npm==0.1.1
optional-django==0.1.0
packaging==24.0
pillow==10.2.0
prompt_toolkit==3.0.47
pycares==4.4.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.8
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
validate-email==1.3
validators==0.22.0
verify-email==2.4.3
vine==5.1.0
wcwidth==0.2.13
asgiref==3.7.2
certifi==2023.7.22
cffi
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django
django-allauth==0.54.0
django-cors-headers
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt
idna
oauthlib
pycparser
PyJWT
python3-openid
pytz
requests
requests-oauthlib
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata
urllib3==2.0.4
